requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-environ==0.11.2
mysqlclient==2.2.6
